Conflicts:
	src/main/webapp/WEB-INF/mybatisAlias/mybatisAlias.xml
	src/main/webapp/WEB-INF/views/tiles/aside.jsp

+++ src/main/java/kr/or/ddit/domain/StudentLecture.java
@@ -7,7 +7,7 @@ public class StudentLecture {
 
 	private int stuNo;
 	private int lecaCd;
-	private float slScore;
+	private Double slScore;
 	private int wsSn;	//장바구니 정렬 순서
 	private int slSn;	//수강신청 정렬 순서
